Fix indeterministic test
The random API_KEY and ID used for the Surveys mock had a chance
of including a special character. If this character was a '#' it
would turn the rest of the URL into a fragment including other
query string paramaters such as 'per_page'

Now explicitly setting 'special_chars=False' rather than assuming
that to be the default.