Update base manager to take access url
The api url for survey monkey can be different based on where you
signed up for your account, see:
https://developer.surveymonkey.com/api/v3/#access-url

This change allows the base manager to take both a connection object with an
access url and the previous connection value by itself. This allows the
api url to be overridden when necessary.